% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet-package.R
\docType{package}
\name{wet-package}
\alias{wet}
\alias{wet-package}
\title{wet: Monthly, Seasonal and Scenario Stream Discharge for the BC Freshwater Atlas}
\description{
Per-segment stream discharge for the British Columbia Freshwater Atlas (FWA). Aggregates Pacific Climate Impacts Consortium (PCIC) VIC-GL gridded runoff and baseflow to FWA fundamental watersheds, accumulates them upstream by area weighting, and reports mean annual and monthly discharge per stream segment for historical and climate-scenario runs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NewGraphEnvironment/wet}
  \item \url{https://newgraphenvironment.github.io/wet/}
  \item Report bugs at \url{https://github.com/NewGraphEnvironment/wet/issues}
}

}
\author{
\strong{Maintainer}: Allan Irvine \email{al@newgraphenvironment.com} (\href{https://orcid.org/0000-0002-3495-2128}{ORCID})

Other contributors:
\itemize{
  \item New Graph Environment Ltd. [copyright holder]
}

}
\keyword{internal}
