% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet-package.R
\docType{package}
\name{wet-package}
\alias{wet}
\alias{wet-package}
\title{wet: Stream Flow and Water Temperature for the BC Freshwater Atlas}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Per-segment stream discharge for the British Columbia Freshwater Atlas (FWA), and flow and water-temperature departures at hydrometric stations. Gridded runoff, from Pacific Climate Impacts Consortium (PCIC) VIC-GL or from an open water balance fitted at Water Survey of Canada (HYDAT) gauges, is sampled to FWA fundamental watersheds and accumulated upstream by area weighting, giving mean annual and monthly discharge per stream segment. Station daily flow and water temperature are summarised per year over month-day windows for departure statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NewGraphEnvironment/wet}
  \item \url{https://newgraphenvironment.github.io/wet/}
  \item Report bugs at \url{https://github.com/NewGraphEnvironment/wet/issues}
}

}
\author{
\strong{Maintainer}: Allan Irvine \email{al@newgraphenvironment.com} (\href{https://orcid.org/0000-0002-3495-2128}{ORCID})

Other contributors:
\itemize{
  \item New Graph Environment Ltd. [copyright holder]
}

}
\keyword{internal}
