% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wet_mm_to_m3s.R
\name{wet_mm_to_m3s}
\alias{wet_mm_to_m3s}
\title{Convert a mean annual depth over an area to discharge}
\usage{
wet_mm_to_m3s(mm, area_m2)
}
\arguments{
\item{mm}{Mean annual depth, mm/yr.}

\item{area_m2}{Contributing area, m^2.}
}
\value{
Discharge, m^3/s.
}
\description{
\code{m3/s = mm/yr * area_m2 / 1000 / (365 * 86400)}, the conversion fwapg uses
(a 365-day year).
}
\examples{
wet_mm_to_m3s(500, 1e8)  # 500 mm/yr over 100 km2 is about 1.59 m3/s
}
